#78 Move paginator methods to trait to be able to reuse it again
